+++ lib/src/annotations/annotations.dart
@@ -1,5 +1,21 @@
+/// ****************************************************
+///                                                    *
+///               By: JhonaCode                        *
+///               Web:     https://jhonacode.com       *
+///               Email:   jhonacode2020@gmail.com     *
+///               Twitter: @jhonacode                  *
+///               Facebook: @jhonacode                 *
+///               Telegram: @jhonacode                 *
+///               March 2023                           *
+///                                                    *
+///         Licensed under the MIT License             *
+///                                                    *
+/// ****************************************************
+
+
 /// Annotations section
 /// Define [Controller]
+
 export 's_controller.dart';
 
 /// Define [Methods]