+++ lib/src/annotations/s_get_mapping.dart
@@ -1,5 +1,19 @@
+/// ****************************************************
+///                                                    *
+///               By: JhonaCode                        *
+///               Web:     https://jhonacode.com       *
+///               Email:   jhonacode2020@gmail.com     *
+///               Twitter: @jhonacode                  *
+///               Facebook: @jhonacode                 *
+///               Telegram: @jhonacode                 *
+///               March 2023                           *
+///                                                    *
+///         Licensed under the MIT License             *
+///                                                    *
+/// ****************************************************
+
+/// For GET request
 class GetMapping {
   final String path;
-  final Symbol methodName;
-  const GetMapping({required this.path, required this.methodName});
+  const GetMapping({required this.path});
 }