+++ lib/src/annotations/s_path_variable.dart
@@ -1,3 +1,18 @@
+/// ****************************************************
+///                                                    *
+///               By: JhonaCode                        *
+///               Web:     https://jhonacode.com       *
+///               Email:   jhonacode2020@gmail.com     *
+///               Twitter: @jhonacode                  *
+///               Facebook: @jhonacode                 *
+///               Telegram: @jhonacode                 *
+///               March 2023                           *
+///                                                    *
+///         Licensed under the MIT License             *
+///                                                    *
+/// ****************************************************
+
+/// Use for define simple data on GET, DELETE request
 class PathVariable {
   final String name;
   const PathVariable(this.name);