+++ lib/src/utils/s_enviroment_data.dart
@@ -19,37 +19,19 @@ import 'package:simple_rest/simple_rest.dart';
 
 /// This class is responsible for reading data from the environment file.
 /// We can define our .env file path or use the default one in the library.
-/// Environment variables:
-/// IP_SERVER=0.0.0.0
-/// PORT_SERVER=8090
-/// JWT_LOCAL_KEY=9832yr794yw
-/// URL_API=https://link your res api [supabase, firebase, etc]
-/// API_KEY
-/// API_AUTH
-/// USER_DB
-/// PASS_DB
-/// URL_DB
+
 class SEnviromentData{
 
-  static String path = '';
+  /// Just define your own path for .env file
+  String path = '';
 
-  static final _enviroment = _loadEnvFile;
+  /// call env value for use on your application
+  dynamic eniValue({required String key}){
+    return _loadEnvFile[key];
+  }
 
-
-
-
-  static String urlApi({String? key})   => _enviroment[key ?? 'URL_API']?.toString()    ?? '127.0.0.1';
-  static String someData({required String key})   => _enviroment[key]?.toString() ?? '';
-  static String apiKey({String? key})   => _enviroment[key ?? 'API_KEY']?.toString()    ?? '';
-  static String apiAuth({String? key})  => _enviroment[key ?? 'API_AUTH']?.toString()   ?? '';
-  static String userDb({String? key})   => _enviroment[key ?? 'USER_DB']?.toString()    ?? '';
-  static String passDb({String? key})   => _enviroment[key ?? 'PASS_DB']?.toString()    ?? '';
-  static String urlDb({String? key})    => _enviroment[key ?? 'URL_DB']?.toString()     ?? '';
-  static String ipServer({String? key}) => _enviroment[ key ?? 'IP_SERVER']?.toString() ?? '0.0.0.0';
-  static String jwtSecretKey ({String? key}) => _enviroment[ key ?? 'JWT_LOCAL_KEY']?.toString() ?? '';
-  static int portServer ({String? key}) => int.tryParse(_enviroment[ key ?? 'PORT_SERVER']?.toString() ?? '8090') ?? 8090;
-
- static Map<String, String> get _loadEnvFile {
+  /// hidden function for return env map
+  Map<String, String> get _loadEnvFile {
 
    String filePath = "../.env";
    final envMap = <String, String>{};