+++ CHANGELOG.md
@@ -6,3 +6,15 @@
 
 ## 0.3.6
 - Update images
+## 0.3.7
+- Update flutter version
+## 0.3.8
+- Route class bug solve
+## 0.3.9
+- Delete JWT library support
+## 0.4.0
+* [BREAKING CHANGES]
+- Use anotation for define Controller.
+- User Shelft library for routes and endpoints.
+- Cors configuration disponible.
+- More